requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
jinja2
python-multipart
aiofiles
cloudscraper
pyyaml
beautifulsoup4
httpx
playwright
